Trying to fix octtree, import issues
