+++ src/config/config_zoning.ini
@@ -1,0 +1,40 @@
+[Input]
+databaseString:None
+database:arcgis
+user:postgres
+password:postgres
+host:None
+port:None
+resolution:100
+EPSGspatialReference:3035
+
+[Sql]
+x:x_mp_100m
+y:y_mp_100m
+value:"Einwohner"
+table:public."population"
+
+[Parameters]
+solve_iteratively:False
+population_threshold:2750
+desired_num_zones:5000
+tolerance:0.1
+lower_population_threshold:1000
+upper_population_threshold:10000
+minimum_zone_population:50
+minimum_zone_area:5000
+
+[Regions]
+EPSGspatialReference:31468
+filename:data/temp/regions_with_land_use
+
+
+[Land Use]
+calculate_land_use:False
+EPSGspatialReference:31468
+filename:../TN_7_Landkreise_Stadt_Muenchen_TUM_Herrn
+class_field:OBJART
+
+[Output]
+filename:zones
+
