+++ src/zoning_algorithm/octtree.py
@@ -1,10 +1,7 @@
-import numpy
-import src.util
 from Queue import Queue
 from rasterstats import zonal_stats
-import octtree
-from collections import defaultdict
 from shapely.geometry import shape
+from helper_functions import *
 
 class Octtree:
     fid_counter = 0
@@ -99,7 +96,7 @@ class OcttreeNode(Octtree):
         return self.count()
 
 def build_out_nodes(Config, region_node, regions, array, affine, pop_threshold):
-    octtree.fid_counter = 0
+    Octtree.fid_counter = 0
 
     octtree_top =  build(region_node.polygon, region_node, array, affine, pop_threshold)
     print "\toriginal number zones: ", octtree_top.count_populated()
@@ -123,7 +120,7 @@ def build(box, parent_node, array, affine, pop_threshold): #list of bottom nodes
     else:  #if np.sum(array) >= pop_threshold and array.size >= 4: # leaf
         #split box into 4
         #recurse to leafs for each subpolygon
-        sub_polygons = util.quarter_polygon(box)
+        sub_polygons = quarter_polygon(box)
         #maybe use clipped and masked sub array
         node = OcttreeNode(box, None, parent_node)
 
@@ -138,7 +135,7 @@ def splice(Config, tree, regions, pop_array, transform):
     region_results = []
     nodes_to_delete = set()
 
-    bounding_area = util.get_region_boundary(regions) #need to check against boundary too.
+    bounding_area = get_region_boundary(regions) #need to check against boundary too.
 
     for region in regions:
         region_results.append({'region':region, 'all':set(), 'to_merge':set()})
@@ -161,14 +158,14 @@ def splice(Config, tree, regions, pop_array, transform):
 
                         intersection = child.polygon.intersection(region_poly) #Check that intersection is a polygon
 
-                        intersections_list = util.get_geom_parts(intersection)
+                        intersections_list = get_geom_parts(intersection)
 
                         for intersection in intersections_list:
                             spliced_node = OcttreeLeaf(intersection, top)
                             spliced_node.region = region
                             region_results[-1]['all'].add(spliced_node)
                             #calculate new population value
-                            spliced_node.value = util.calculate_pop_value(spliced_node, pop_array, transform)
+                            spliced_node.value = calculate_pop_value(spliced_node, pop_array, transform)
                             if spliced_node.is_acceptable(Config):
                                 top.children.append(spliced_node)
                             else:
@@ -203,7 +200,7 @@ def merge(Config, region_results):
                 if node not in node.parent.children:
                     node.parent.children.append(node)
             else:
-                best_neighbour = util.find_best_neighbour(node, region_nodes)
+                best_neighbour = find_best_neighbour(node, region_nodes)
                 if best_neighbour:
                     best_neighbour.polygon = best_neighbour.polygon.union(node.polygon)
                     best_neighbour.value = best_neighbour.value + node.value
