+++ src/zoning_algorithm/population_processing.py
@@ -1,9 +1,9 @@
 import sys, os
-import util, octtree
+import src.util as util
+import octtree
 import ConfigParser
-from shapely.geometry import shape, Polygon
-from shapely.ops import cascaded_union
 from fiona.crs import from_epsg
+import rasterio
 
 Config = ConfigParser.ConfigParser(allow_no_value=True)
 
@@ -11,42 +11,21 @@ if len(sys.argv) == 1 or not os.path.exists(sys.argv[1]):
     raise IOError("please supply a configuration file as a program arugment")
 Config.read(sys.argv[1])
 
-#next step, find the 'power of two' box that best captures the polygonal boundary area.
-resolution = Config.getint("Input", "resolution")
-zonesSaptialRef = from_epsg(Config.getint("Input", "EPSGspatialReference"))
-regions_file = Config.get("Regions", "filename")
+with rasterio.open(Config.get("Input", "raster")) as r:
+    raster_array = r.read(1)
+    transform = r.affine
+    zonesSaptialRef = r.crs
 
-regions = util.load_regions(regions_file, zonesSaptialRef)
+regions = util.load_regions(Config)
 boundary = util.get_region_boundary(regions)
 
-(min_x, min_y, max_x, max_y) = map(int, boundary.bounds) #given boundary, get envelope of polygon, as integers
-print (min_x, min_y, max_x, max_y)
-
-
-max_dimension = max(max_x - min_x, max_y - min_y)
-sub_array_size = util.next_power_of_2(max_dimension / resolution)
-print "array will be size: ", sub_array_size
-
-#use centroid of area to work out array origins
-array_origin_x = (max_x + min_x - sub_array_size*resolution)/ 2
-array_origin_y = (max_y + min_y - sub_array_size*resolution)/ 2
-
-(pop_array, transform) = util.load_data(Config, array_origin_x, array_origin_y, sub_array_size)
-
-
-
-boundary_box = Polygon([transform*(0, 0),
-                        transform*(0, sub_array_size),
-                        transform*(sub_array_size, sub_array_size),
-                        transform*(sub_array_size, 0)])
-
-region_octtree = octtree.OcttreeNode(boundary_box, None, None)
+region_octtree = octtree.OcttreeNode(boundary.envelope, None, None)
 
 if Config.getboolean("Parameters", "solve_iteratively"):
-    region_octtree = util.solve_iteratively(Config, region_octtree, regions, pop_array, transform)
+    region_octtree = util.solve_iteratively(Config, region_octtree, regions, raster_array, transform)
 else:
     pop_threshold =  Config.getint("Parameters", "population_threshold")
-    region_octtree = octtree.build_out_nodes(Config, region_octtree, regions, pop_array, transform, pop_threshold)
+    region_octtree = octtree.build_out_nodes(Config, region_octtree, regions, raster_array, transform, pop_threshold)
 
 
 shapefile = Config.get("Land Use", "filename")
